LayoutTests/accessibility/mac/search-text/search-text-expected.txt

The quick brown fox jumps over the lazy dog.

TEXT2: The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.

This tests the ability to search for text given a starting point and a direction for the search.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'lazy'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'quick'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'the'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'The'
PASS result.length is 3
PASS body.stringForTextMarkerRange(result[i]) is 'dog'
PASS body.stringForTextMarkerRange(result[i]) is 'dog'
PASS body.stringForTextMarkerRange(result[i]) is 'dog'
PASS result.length is 3
PASS body.stringForTextMarkerRange(result[i]) is 'fox'
PASS body.stringForTextMarkerRange(result[i]) is 'fox'
PASS body.stringForTextMarkerRange(result[i]) is 'fox'
PASS typeof result is 'undefined'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'over'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'fox'
PASS result.length is 1
PASS text.stringForTextMarkerRange(result[0]) is 'fox'
PASS successfullyParsed is true

TEST COMPLETE
